Split types into msg and state
