+++ contracts/native-staking/src/state.rs
@@ -1,8 +1,6 @@
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::Addr;
 
-/*** state ***/
-
 #[cw_serde]
 pub struct Config {
     /// The denom we accept for staking
@@ -13,24 +11,4 @@ pub struct Config {
 
     /// The address of the vault contract (where we get and return stake)
     pub vault: Addr,
-}
-
-/**** api ****/
-
-pub type ConfigResponse = Config;
-
-#[cw_serde]
-pub struct ProxyByOwnerResponse {
-    pub proxy: String,
-}
-
-#[cw_serde]
-pub struct OwnerByProxyResponse {
-    pub owner: String,
-}
-
-/// The message that is binary encoded in `receive_stake(..msg)`
-#[cw_serde]
-pub struct StakeMsg {
-    pub validator: String,
 }
